backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
google-genai>=1.0.0
yt-dlp>=2025.1.1
